requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.9
pytest==8.1.1
flake8==7.0.0
black==24.3.0
httpx==0.23.0
strawberry-graphql[fastapi]==0.216.1
